rename crap

+++ src/_commands/unlink.js
@@ -1,8 +1,4 @@
-const Discord = require(`discord.js`);
 const User = require(`../models/user.model`);
-const {
-    config
-} = require(`../index.js`);
 
 module.exports = {
     name: `unlink`,
@@ -11,16 +7,12 @@ module.exports = {
 };
 
 module.exports.run = async (client, message, args) => {
-    if (message.author.id == `448619102198693889`) return;
+    if (message.author.id === `448619102198693889`) return;
 
-    const account = await User.findOne({
-        discordID: message.author.id
-    });
+    const account = await User.findOne({ discordID: message.author.id });
     if (!account) return message.channel.send(`${message.author} You are not linked to an account!`);
 
-    User.deleteOne({
-        discordID: message.author.id
-    })
-        .catch(err => message.channel.send(`${message.author} Failed to unlink your account. Please contact a developer.`))
+    User.deleteOne({ discordID: message.author.id })
+        .catch(() => message.channel.send(`${message.author} Failed to unlink your account. Please contact a developer.`))
         .then(() => message.channel.send(`${message.author} Succesfully unlinked your account.`));
 };
